:construction: Play a card
