+++ src/room/room.websocket.gateway.ts
@@ -12,13 +12,15 @@ import {HttpException} from "@nestjs/common/exceptions";
 import {RoomService} from "./service/room.service";
 import {Message} from "./dto/room.dto";
 import { jwtDecode } from "jwt-decode";
-import {Pli, Round} from "./room.model";
+import {Play, PlayCard, Pli, Round} from "./room.model";
+import {GameService} from "../game/service/game.service";
 
 @WebSocketGateway({ cors : '*', namespace: 'room'})
 export class RoomWebsocketGateway implements OnGatewayConnection, OnGatewayDisconnect {
 
   constructor(private readonly redisService: RedisService,
-              private readonly roomService: RoomService) {}
+              private readonly roomService: RoomService,
+              private readonly gameService: GameService) {}
 
   @WebSocketServer() server;
 
@@ -72,7 +74,7 @@ export class RoomWebsocketGateway implements OnGatewayConnection, OnGatewayDisco
     console.log(`API - ${client.data.user.username} is starting ${slug}`)
     try {
       if (await this.redisService.exists(`room:${slug}`)) {
-        this.roomService.startGame(slug, client.data.user).then((users) => {
+        this.gameService.startGame(slug, client.data.user).then((users) => {
           for (const user of users) {
             this.server.to(user.socketId).emit('cards', user.cards);
           }
@@ -97,7 +99,7 @@ export class RoomWebsocketGateway implements OnGatewayConnection, OnGatewayDisco
   async newRound(@ConnectedSocket() client: Socket, @MessageBody() body : { slug: string, nbCards: number }): Promise<{}> {
     try {
       if (await this.redisService.exists(`room:${body.slug}`)) {
-        this.roomService.newRound(body).then((users) => {
+        this.gameService.newRound(body).then((users) => {
           for (const user of users) {
             this.server.to(user.socketId).emit('cards', user.cards);
           }
@@ -122,7 +124,7 @@ export class RoomWebsocketGateway implements OnGatewayConnection, OnGatewayDisco
   async endRound(@ConnectedSocket() client: Socket, @MessageBody() round : Round): Promise<{}> {
     try {
       if (await this.redisService.exists(`room:${round.slug}`)) {
-        this.roomService.endRound(round).then((users) => {
+        this.gameService.endRound(round).then((users) => {
           this.server.to(round.slug).emit('endRound', users); // broadcast messages endRound
         })
         .catch((e) => {
@@ -144,7 +146,7 @@ export class RoomWebsocketGateway implements OnGatewayConnection, OnGatewayDisco
   async newPli(@ConnectedSocket() client: Socket, @MessageBody() pli : Pli): Promise<{}> {
     try {
       if (await this.redisService.exists(`room:${pli.slug}`)) {
-        this.roomService.newPli(pli).then(([winner, bonus]) => {
+        this.gameService.newPli(pli).then(([winner, bonus]) => {
           this.server.to(pli.slug).emit('newPli', [winner, bonus]); // broadcast messages newPli
         })
         .catch((e) => {
@@ -161,4 +163,36 @@ export class RoomWebsocketGateway implements OnGatewayConnection, OnGatewayDisco
       }
     }
   }
+
+  @SubscribeMessage('play')
+  async play(@ConnectedSocket() client: Socket, @MessageBody() playcard : PlayCard): Promise<{}> {
+    try {
+      if (await this.redisService.exists(`room:${playcard.slug}`)) {
+        this.gameService.playCard(playcard).then((playcard) => {
+          this.server.to(playcard.slug).emit('playcard', playcard); // broadcast messages playcard
+        })
+          .catch((e) => {
+            console.log(e)
+            throw new HttpException(e.message, e.status);
+          });
+      } else {
+        throw new HttpException("La room n'existe pas", 404);
+      }
+    } catch (e) {
+      return {
+        message: e.message,
+        status: e.status,
+      }
+    }
+  }
 }
+
+// ===> CHAQUE CARTE JOUER
+// Get card + user
+// Check if user has the card
+// append card + user dans room:slug:nbManche:pli
+// emit carte joué
+
+// si room:slug:nbManche:pli.length == room.users.length
+//   whoWinTheTrick(room:slug:nbManche:pli)
+// <====
